Use On-demand proximity sensor state for dt2w
dt2w is detected and right after it scheduled for the power action, check the
proximity sensor if it can be woken up.
If the sensor is covered, ignore the dt2w event, otherwise continue to wake up
the device.

Additional debugs added to assist any issues with dt2w

Change-Id: Ifc25c87f0853d776f2e01772bc1bdb9cfc82b3ef