Reserve the hardboot page correctly
The location 0x8c300000 is actually already allocated, so changes to ramoops are
of no use. This is revealed when i started dumping the memory block statistics.
So had to push the peripheral memory region 1MB up, so i have a vacant 1MB
region.

With this change, the first time i opted to boot secondary ROM resulted into
Mi-Logo stuck. There seems to be some race-condition/dead-lock/name-it, as i
have not yet looked into fully, but have captured the log.
So a long-press-power-button, actually triggered the boot again, and Viola it
actually booted into secondary ROM. And a reboot from there, in-comes the system
with Primary ROM.

Okay, so, there are few glitches, but there is a progress on kexec patch. Let us
see if we can fix those bitchy glitches soon.

Change-Id: Ia57625573d12bee2002766b9e6e926441986b897